requirements.txt
# Connect4 RL Training System - Dependencies
# Version: 1.0.0

# Core libraries
numpy>=1.24.0
torch>=2.0.0
gymnasium>=0.29.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0

# Optional for notebooks
matplotlib>=3.7.0
jupyter>=1.0.0

# Development tools
pytest>=7.0.0
black>=23.0.0
mypy>=1.0.0
flake8>=6.0.0
isort>=5.12.0